requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.31.0
langchain-groq==0.2.1
langchain==0.3.28
pypdf==4.2.0
networkx==3.3
python-multipart==0.0.9
